Initial commit of Agave Tequilana
This is a re-implementation of Tequila War / Cactus, a scoring add-on for
VGA Planets.

This initial commit provides a baseline to improve on, and supports all
features of the original implementation with hopefully none of the bugs.
The only extra feature is multi-language support using PHost's regular
interface, as well as some command-line options.

The source code is structured similar to Starbase Reloaded, and re-uses some
of its functionality.